[Ref]: NowPlaying 로직 계층 코드 분리
